+++ pvzclass/Classes.hpp
@@ -1,4 +1,6 @@
 ﻿#pragma once
+#include "Classes/ZenGarden.hpp"
+
 #include "Classes/Definition.hpp"
 
 #include "Classes/ChallengeScreen.hpp"
