+++ pvzclass/PVZ.cpp
@@ -146,11 +146,6 @@ PVZ::Board PVZ::GetBoard()
 	return(address == 0 ? INVALID_BASEADDRESS : Board(address));
 }
 
-PVZ::ZenGarden PVZ::GetZenGarden()
-{
-	return ZenGarden(Memory::ReadPointer(0x6A9EC0, 0x81C));
-}
-
 PVZ::SaveData PVZ::GetSaveData()
 {
 	return SaveData(Memory::ReadPointer(0x6A9EC0, 0x82C));
