+++ pvzclass/PVZ.h
@@ -1061,18 +1061,6 @@ namespace PVZ
 	};
 	//if anyone want a class for calling functions in bass.dll to totally control the music in game,just tell me
 
-	/// @brief 禅境花园类，包含部分与禅境花园有关的函数。
-	class ZenGarden : public BaseClass
-	{
-	public:
-		ZenGarden(int address) : BaseClass(address) {};
-		PVZApp GetLawnApp();
-		PVZ::Board GetBoard();
-		T_PROPERTY(GardenScene::GardenScene, GardenType, __get_GardenType, __set_GardenType, 0x8);
-		bool IsFull(bool consider_items);
-		Snail GetSnail();
-	};
-
 #pragma endregion
 
 #pragma region methods
@@ -1083,7 +1071,6 @@ namespace PVZ
 	Board GetBoard();
 	SaveData GetSaveData();
 	Music GetMusic();
-	ZenGarden GetZenGarden();
 
 #pragma endregion
